split logger and memory
