+++ src/smolagents/agents.py
@@ -26,7 +26,11 @@ from rich.rule import Rule
 from rich.syntax import Syntax
 from rich.text import Text
 
-from smolagents.logger import ActionStep, PlanningStep, SystemPromptStep, TaskStep, ToolCall
+from smolagents.logger import (
+    AgentLogger,
+    LogLevel,
+)
+from smolagents.memory import ActionStep, AgentMemory, PlanningStep, SystemPromptStep, TaskStep, ToolCall
 from smolagents.types import AgentAudio, AgentImage, handle_agent_output_types
 from smolagents.utils import (
     AgentError,
@@ -45,10 +49,6 @@ from .local_python_executor import (
     BASE_BUILTIN_MODULES,
     LocalPythonInterpreter,
     fix_final_answer_code,
-)
-from .logger import (
-    AgentLogger,
-    LogLevel,
 )
 from .models import (
     ChatMessage,
@@ -188,6 +188,7 @@ class MultiStepAgent:
         self.system_prompt = self.initialize_system_prompt()
         self.input_messages = None
         self.task = None
+        self.memory = AgentMemory()
         self.logger = AgentLogger(level=verbosity_level)
         self.monitor = Monitor(self.model, self.logger)
         self.step_callbacks = step_callbacks if step_callbacks is not None else []
@@ -213,7 +214,7 @@ class MultiStepAgent:
         the LLM.
         """
         memory = []
-        for step_log in self.logger.steps:
+        for step_log in self.memory.steps:
             memory.extend(step_log.to_messages(summary_mode=summary_mode))
         return memory
 
@@ -290,7 +291,7 @@ class MultiStepAgent:
             ]
         try:
             chat_message: ChatMessage = self.model(self.input_messages)
-            self.logger.log_chat_messages(chat_message)
+            self.memory.log_chat_messages(chat_message)
             return chat_message.content
         except Exception as e:
             return f"Error in generating final LLM output:\n{e}"
@@ -385,11 +386,11 @@ You have been provided with these additional arguments, that you can access usin
         system_prompt_step = SystemPromptStep(system_prompt=self.system_prompt)
 
         if reset:
-            self.logger.reset()
-            self.logger.log_step(system_prompt_step, position=0)
+            self.memory.reset()
+            self.memory.log_step(system_prompt_step, position=0)
             self.monitor.reset()
         else:
-            self.logger.log_step(system_prompt_step, position=0)
+            self.memory.log_step(system_prompt_step, position=0)
 
         self.logger.log(
             Panel(
@@ -402,7 +403,7 @@ You have been provided with these additional arguments, that you can access usin
             level=LogLevel.INFO,
         )
 
-        self.logger.log_step(TaskStep(task=self.task, task_images=images))
+        self.memory.log_step(TaskStep(task=self.task, task_images=images))
         if single_step:
             step_start_time = time.time()
             step_log = ActionStep(start_time=step_start_time, observations_images=images)
@@ -459,7 +460,7 @@ You have been provided with these additional arguments, that you can access usin
             finally:
                 step_log.end_time = time.time()
                 step_log.duration = step_log.end_time - step_start_time
-                self.logger.log_step(step_log)
+                self.memory.log_step(step_log)
                 for callback in self.step_callbacks:
                     # For compatibility with old callbacks that don't take the agent as an argument
                     if len(inspect.signature(callback).parameters) == 1:
@@ -480,7 +481,7 @@ You have been provided with these additional arguments, that you can access usin
             final_step_log.action_output = final_answer
             final_step_log.end_time = time.time()
             final_step_log.duration = step_log.end_time - step_start_time
-            self.logger.log_step(final_step_log)
+            self.memory.log_step(final_step_log)
             for callback in self.step_callbacks:
                 # For compatibility with old callbacks that don't take the agent as an argument
                 if len(inspect.signature(callback).parameters) == 1:
@@ -515,7 +516,7 @@ Now begin!""",
             }
 
             chat_message: ChatMessage = self.model([message_prompt_facts, message_prompt_task])
-            self.logger.log_chat_messages(chat_message)
+            self.memory.log_chat_messages(chat_message)
             answer_facts = chat_message.content
 
             message_system_prompt_plan = {
@@ -535,7 +536,7 @@ Now begin!""",
                 [message_system_prompt_plan, message_user_prompt_plan],
                 stop_sequences=["<end_plan>"],
             )
-            self.logger.log_chat_messages(chat_message_plan)
+            self.memory.log_chat_messages(chat_message_plan)
             answer_plan = chat_message_plan.content
 
             final_plan_redaction = f"""Here is the plan of action that I will follow to solve the task:
@@ -546,7 +547,7 @@ Now begin!""",
 ```
 {answer_facts}
 ```""".strip()
-            self.logger.log_step(PlanningStep(plan=final_plan_redaction, facts=final_facts_redaction))
+            self.memory.log_step(PlanningStep(plan=final_plan_redaction, facts=final_facts_redaction))
             self.logger.log(
                 Rule("[bold]Initial plan", style="orange"),
                 Text(final_plan_redaction),
@@ -569,7 +570,7 @@ Now begin!""",
             chat_message: ChatMessage = self.model(
                 [facts_update_system_prompt] + agent_memory + [facts_update_message]
             )
-            self.logger.log_chat_messages(chat_message)
+            self.memory.log_chat_messages(chat_message)
             facts_update = chat_message.content
 
             # Redact updated plan
@@ -591,7 +592,7 @@ Now begin!""",
                 [plan_update_message] + agent_memory + [plan_update_message_user],
                 stop_sequences=["<end_plan>"],
             )
-            self.logger.log_chat_messages(chat_message)
+            self.memory.log_chat_messages(chat_message)
             plan_update = chat_message.content
 
             # Log final facts and plan
@@ -600,7 +601,7 @@ Now begin!""",
 ```
 {facts_update}
 ```"""
-            self.logger.log_step(PlanningStep(plan=final_plan_redaction, facts=final_facts_redaction))
+            self.memory.log_step(PlanningStep(plan=final_plan_redaction, facts=final_facts_redaction))
             self.logger.log(
                 Rule("[bold]Updated plan", style="orange"),
                 Text(final_plan_redaction),
@@ -657,7 +658,7 @@ class ToolCallingAgent(MultiStepAgent):
                 tools_to_call_from=list(self.tools.values()),
                 stop_sequences=["Observation:"],
             )
-            self.logger.log_chat_messages(model_message)
+            self.memory.log_chat_messages(model_message)
             if model_message.tool_calls is None or len(model_message.tool_calls) == 0:
                 raise Exception("Model did not call any tools. Call `final_answer` tool to return a final answer.")
             tool_call = model_message.tool_calls[0]
@@ -822,7 +823,7 @@ class CodeAgent(MultiStepAgent):
                 stop_sequences=["<end_code>", "Observation:"],
                 **additional_args,
             )
-            self.logger.log_chat_messages(chat_message)
+            self.memory.log_chat_messages(chat_message)
             llm_output = chat_message.content
             log_entry.llm_output = llm_output
         except Exception as e:
@@ -856,7 +857,7 @@ class CodeAgent(MultiStepAgent):
             ToolCall(
                 name="python_interpreter",
                 arguments=code_action,
-                id=f"call_{len(self.logger.steps)}",
+                id=f"call_{len(self.memory.steps)}",
             )
         ]
 
@@ -968,9 +969,4 @@ class ManagedAgent:
             return output
 
 
-__all__ = [
-    "ManagedAgent",
-    "MultiStepAgent",
-    "CodeAgent",
-    "ToolCallingAgent",
-]
+__all__ = ["ManagedAgent", "MultiStepAgent", "CodeAgent", "ToolCallingAgent", "AgentMemory"]
