+++ src/smolagents/logger.py
@@ -1,180 +1,14 @@
 import json
-from dataclasses import asdict, dataclass
 from enum import IntEnum
-from typing import TYPE_CHECKING, Any, Dict, List
+from typing import TYPE_CHECKING
 
 from rich.console import Console
 from rich.rule import Rule
 from rich.syntax import Syntax
 
-from smolagents.models import MessageRole
-from smolagents.utils import AgentError, make_json_serializable
-
 
 if TYPE_CHECKING:
-    from smolagents.models import ChatMessage
-
-
-console = Console()
-
-
-@dataclass
-class Message:
-    role: MessageRole
-    content: str | list[dict]
-
-    def dict(self):
-        return asdict(self)
-
-
-@dataclass
-class ToolCall:
-    name: str
-    arguments: Any
-    id: str
-
-    def dict(self):
-        return {
-            "id": self.id,
-            "type": "function",
-            "function": {
-                "name": self.name,
-                "arguments": make_json_serializable(self.arguments),
-            },
-        }
-
-
-class AgentStepLog:
-    raw: Any  # This is a placeholder for the raw data that the agent logs
-
-    def dict(self):
-        return asdict(self)
-
-    def to_messages(self, **kwargs) -> List[Dict[str, Any]]:
-        raise NotImplementedError
-
-
-@dataclass
-class ActionStep(AgentStepLog):
-    agent_memory: List[Dict[str, str]] | None = None
-    tool_calls: List[ToolCall] | None = None
-    start_time: float | None = None
-    end_time: float | None = None
-    step_number: int | None = None
-    error: AgentError | None = None
-    duration: float | None = None
-    llm_output: str | None = None
-    observations: str | None = None
-    observations_images: List[str] | None = None
-    action_output: Any = None
-
-    def dict(self):
-        # We overwrite the method to parse the tool_calls and action_output manually
-        return {
-            "agent_memory": self.agent_memory,
-            "tool_calls": [tc.dict() for tc in self.tool_calls] if self.tool_calls else [],
-            "start_time": self.start_time,
-            "end_time": self.end_time,
-            "step": self.step_number,
-            "error": self.error.dict() if self.error else None,
-            "duration": self.duration,
-            "llm_output": self.llm_output,
-            "observations": self.observations,
-            "action_output": make_json_serializable(self.action_output),
-        }
-
-    def to_messages(self, summary_mode: bool = False, return_memory: bool = False) -> List[Dict[str, Any]]:
-        memory = []
-        if self.agent_memory is not None and return_memory:
-            message = Message(MessageRole.SYSTEM, self.agent_memory)
-            memory.append(message.dict())
-        if self.llm_output is not None and not summary_mode:
-            message = Message(MessageRole.ASSISTANT, [{"type": "text", "text": self.llm_output.strip()}])
-            memory.append(message.dict())
-
-        if self.tool_calls is not None:
-            message = Message(
-                MessageRole.ASSISTANT, [{"type": "text", "text": str([tc.dict() for tc in self.tool_calls])}]
-            )
-            memory.append(message.dict())
-
-        if self.error is not None:
-            message_content = (
-                "Error:\n"
-                + str(self.error)
-                + "\nNow let's retry: take care not to repeat previous errors! If you have retried several times, try a completely different approach.\n"
-            )
-            if self.tool_calls is None:
-                tool_response_message = Message(MessageRole.ASSISTANT, [{"type": "text", "text": message_content}])
-            else:
-                tool_response_message = Message(
-                    MessageRole.TOOL_RESPONSE, f"Call id: {self.tool_calls[0].id}\n{message_content}"
-                )
-
-            memory.append(tool_response_message.dict())
-        else:
-            if self.observations is not None and self.tool_calls is not None:
-                tool_response_message = Message(
-                    MessageRole.TOOL_RESPONSE,
-                    f"Call id: {self.tool_calls[0].id}\nObservation:\n{self.observations}",
-                )
-                memory.append(tool_response_message.dict())
-        if self.observations_images:
-            thought_message_image = Message(
-                MessageRole.USER,
-                [{"type": "text", "text": "Here are the observed images:"}]
-                + [
-                    {
-                        "type": "image",
-                        "image": image,
-                    }
-                    for image in self.observations_images
-                ],
-            )
-            memory.append(thought_message_image.dict())
-        return memory
-
-
-@dataclass
-class PlanningStep(AgentStepLog):
-    plan: str
-    facts: str
-
-    def to_messages(self, summary_mode: bool, **kwargs) -> List[Dict[str, str]]:
-        memory = []
-        thought_message = Message(MessageRole.ASSISTANT, f"[FACTS LIST]:\n{self.facts.strip()}")
-        memory.append(thought_message.dict())
-
-        if not summary_mode:
-            thought_message = Message(MessageRole.ASSISTANT, f"[PLAN]:\n{self.plan.strip()}")
-            memory.append(thought_message.dict())
-        return memory
-
-
-@dataclass
-class TaskStep(AgentStepLog):
-    task: str
-    task_images: List[str] | None = None
-
-    def to_messages(self, summary_mode: bool, **kwargs) -> List[Dict[str, str]]:
-        content = [{"type": "text", "text": f"New task:\n{self.task}"}]
-        if self.task_images:
-            for image in self.task_images:
-                content.append({"type": "image", "image": image})
-
-        message = Message(MessageRole.USER, content)
-        return [message.dict()]
-
-
-@dataclass
-class SystemPromptStep(AgentStepLog):
-    system_prompt: str
-
-    def to_messages(self, summary_mode: bool, **kwargs) -> List[Dict[str, str]]:
-        if not summary_mode:
-            message = Message(MessageRole.SYSTEM, [{"type": "text", "text": self.system_prompt.strip()}])
-            return [message.dict()]
-        return []
+    from smolagents.memory import AgentMemory
 
 
 class LogLevel(IntEnum):
@@ -186,12 +20,7 @@ class LogLevel(IntEnum):
 class AgentLogger:
     def __init__(self, level: LogLevel = LogLevel.INFO):
         self.level = level
-        self.steps: List[ActionStep] = []
-        self.chat_messages: List[ChatMessage] = []
         self.console = Console()
-
-    def reset(self):
-        self.steps = []
 
     def log(self, *args, level: str | LogLevel = LogLevel.INFO, **kwargs):
         """Logs a message to the console.
@@ -204,35 +33,7 @@ class AgentLogger:
         if level <= self.level:
             self.console.print(*args, **kwargs)
 
-    def log_step(self, step: AgentStepLog, position: int = None):
-        """Logs an agent execution step for ulterior processing.
-
-        Args:
-            step (AgentStepLog)
-            position (int, optional): Position at which to insert the item.
-                Defaults to None, in which case item is added to the end of the list.
-                Should only be used to insert the system prompt at the start of the list.
-        """
-        if position is None:
-            self.steps.append(step)
-        else:
-            assert position == 0, "Position should only be 0 for system prompt."
-            self.steps = [step] + self.steps[1:]  # we replace the system prompt
-
-    def log_chat_messages(self, msg: "ChatMessage"):
-        """Logs all raw model outputs throughout the run, for debugging purposes."""
-        self.chat_messages.append(msg)
-
-    def get_succinct_logs(self):
-        return [{key: value for key, value in log.items() if key != "agent_memory"} for log in self.steps]
-
-    def get_extended_logs(self) -> list[dict]:
-        return {
-            "steps": [step.dict() for step in self.steps],
-            "chat_messages": [msg.dict() for msg in self.chat_messages],
-        }
-
-    def replay(self, with_memory: bool = False):
+    def replay(self, agent_memory: "AgentMemory", full: bool = False):
         """Prints a pretty replay of the agent's steps.
 
         Args:
@@ -240,8 +41,8 @@ class AgentLogger:
                 Careful: will increase log length exponentially. Use only for debugging.
         """
         memory = []
-        for step_log in self.steps:
-            memory.extend(step_log.to_messages(return_memory=with_memory, summary_mode=False))
+        for step_log in agent_memory.steps:
+            memory.extend(step_log.to_messages(return_memory=full))
 
         self.console.log("Replaying the agent's steps:")
         ix = 0
