+++ tests/test_agents.py
@@ -307,9 +307,9 @@ class AgentTests(unittest.TestCase):
         output = agent.run("What is 2 multiplied by 3.6452?")
         assert isinstance(output, str)
         assert "7.2904" in output
-        assert agent.logger.steps[1].task == "What is 2 multiplied by 3.6452?"
-        assert "7.2904" in agent.logger.steps[2].observations
-        assert agent.logger.steps[3].llm_output is None
+        assert agent.memory.steps[1].task == "What is 2 multiplied by 3.6452?"
+        assert "7.2904" in agent.memory.steps[2].observations
+        assert agent.memory.steps[3].llm_output is None
 
     def test_toolcalling_agent_handles_image_tool_outputs(self):
         from PIL import Image
@@ -352,8 +352,8 @@ class AgentTests(unittest.TestCase):
         output = agent.run("What is 2 multiplied by 3.6452?")
         assert isinstance(output, float)
         assert output == 7.2904
-        assert agent.logger.steps[1].task == "What is 2 multiplied by 3.6452?"
-        assert agent.logger.steps[3].tool_calls == [
+        assert agent.memory.steps[1].task == "What is 2 multiplied by 3.6452?"
+        assert agent.memory.steps[3].tool_calls == [
             ToolCall(name="python_interpreter", arguments="final_answer(7.2904)", id="call_3")
         ]
 
@@ -370,30 +370,30 @@ class AgentTests(unittest.TestCase):
         agent = CodeAgent(tools=[PythonInterpreterTool()], model=fake_code_model)
         output = agent.run("What is 2 multiplied by 3.6452?", reset=True)
         assert output == 7.2904
-        assert len(agent.logger.steps) == 4
+        assert len(agent.memory.steps) == 4
 
         output = agent.run("What is 2 multiplied by 3.6452?", reset=False)
         assert output == 7.2904
-        assert len(agent.logger.steps) == 6
+        assert len(agent.memory.steps) == 6
 
         output = agent.run("What is 2 multiplied by 3.6452?", reset=True)
         assert output == 7.2904
-        assert len(agent.logger.steps) == 4
+        assert len(agent.memory.steps) == 4
 
     def test_code_agent_code_errors_show_offending_line_and_error(self):
         agent = CodeAgent(tools=[PythonInterpreterTool()], model=fake_code_model_error)
         output = agent.run("What is 2 multiplied by 3.6452?")
         assert isinstance(output, AgentText)
         assert output == "got an error"
-        assert "Code execution failed at line 'error_function()'" in str(agent.logger.steps[2].error)
-        assert "ValueError" in str(agent.logger.steps)
+        assert "Code execution failed at line 'error_function()'" in str(agent.memory.steps[2].error)
+        assert "ValueError" in str(agent.memory.steps)
 
     def test_code_agent_syntax_error_show_offending_lines(self):
         agent = CodeAgent(tools=[PythonInterpreterTool()], model=fake_code_model_syntax_error)
         output = agent.run("What is 2 multiplied by 3.6452?")
         assert isinstance(output, AgentText)
         assert output == "got an error"
-        assert '    print("Failing due to unexpected indent")' in str(agent.logger.steps)
+        assert '    print("Failing due to unexpected indent")' in str(agent.memory.steps)
 
     def test_setup_agent_with_empty_toolbox(self):
         ToolCallingAgent(model=FakeToolCallModel(), tools=[])
@@ -405,8 +405,8 @@ class AgentTests(unittest.TestCase):
             max_steps=5,
         )
         answer = agent.run("What is 2 multiplied by 3.6452?")
-        assert len(agent.logger.steps) == 8
-        assert type(agent.logger.steps[-1].error) is AgentMaxStepsError
+        assert len(agent.memory.steps) == 8
+        assert type(agent.memory.steps[-1].error) is AgentMaxStepsError
         assert isinstance(answer, str)
 
     def test_tool_descriptions_get_baked_in_system_prompt(self):
